Refactor STR page and page container into seperate files
